% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{runApp}
\alias{runApp}
\title{Run the coding app}
\usage{
runApp(use_test_data = FALSE)
}
\arguments{
\item{use_test_data}{(FOR DEVELOPERS ONLY) Do you want to use test data for this app?}
}
\description{
Run the coding app
}
